requirements.txt
streamlit
requests
python-dotenv
crewai==0.1.24
langchain==0.0.354
langchain-core==0.1.5
openai
tiktoken==0.5.1
chromadb
sentence-transformers
